flight-delay-system/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
joblib==1.4.2
python-dotenv==1.0.1
pydantic==2.7.1

